refactor: complete modularization of ConfigurationValidationRules
Split 1429-line monolith into focused modules:
- TripodValidationRules.ts (4 rules, 230 lines)
- QuadVeeValidationRules.ts (4 rules, 260 lines)
- OmniMechValidationRules.ts (3 rules, 130 lines)
- BipedValidationRules.ts (1 rule, 65 lines)
- GenericValidationRules.ts (1 rule, 60 lines)

ConfigurationValidationRules.ts now 52 lines (aggregator only)
Total: 1429 → 1027 lines (28% reduction) across 8 focused files
